+++ src/gui/gtk3/rwreg_store.rs
@@ -1,4 +1,4 @@
-use crate::platine::BoxedPlatine;
+use crate::{registers::Register, platine::BoxedPlatine};
 use gtk::prelude::*;
 
 /// GtkTreestore for a Rwreg
@@ -33,7 +33,7 @@ impl RwregStore {
                 None,
                 &[0, 1, 2, 3, 4],
                 &[
-                    &reg.rwreg_nr(),
+                    &reg.reg_nr(),
                     &reg.range(),
                     &reg.value(),
                     &reg.description(),